ver.0.5
カテゴリー削除・メニューのカテゴリー反映

+++ ekchuah_cms/menu/menu_input.php
@@ -16,7 +16,10 @@ if(isset($_GET['back'])){
     $menu_price = $_SESSION['menu_price'];
     $category = $_SESSION['category'];
     $endDate = $_SESSION['endDate'];
-    unlink('tmp/'.$_SESSION['menu_id'].'.jpg');
+    if(file_exists('tmp/'.$_SESSION['menu_id'].'.jpg')){
+        unlink('tmp/'.$_SESSION['menu_id'].'.jpg');
+    }
+    
 }
 
 //selectにカテゴリーを表示
